Refactor CLI progress module into dedicated components
