+++ crates/cli/src/lib.rs
@@ -98,7 +98,7 @@ use std::process::{Child, Command, Stdio};
 use std::str::FromStr;
 use std::sync::mpsc;
 use std::thread;
-use std::time::{Duration, SystemTime};
+use std::time::SystemTime;
 
 #[cfg(unix)]
 use std::os::unix::{ffi::OsStringExt, fs::PermissionsExt};
@@ -112,10 +112,10 @@ use rsync_core::{
     client::{
         AddressMode, BandwidthLimit, BindAddress, ClientConfig, ClientEntryKind,
         ClientEntryMetadata, ClientEvent, ClientEventKind, ClientOutcome, ClientProgressObserver,
-        ClientProgressUpdate, ClientSummary, CompressionSetting, DeleteMode, DirMergeEnforcedKind,
-        DirMergeOptions, FilterRuleKind, FilterRuleSpec, HumanReadableMode, ModuleListOptions,
-        ModuleListRequest, RemoteFallbackArgs, RemoteFallbackContext, StrongChecksumChoice,
-        TransferTimeout, parse_skip_compress_list, run_client_or_fallback,
+        CompressionSetting, DeleteMode, DirMergeEnforcedKind, DirMergeOptions, FilterRuleKind,
+        FilterRuleSpec, HumanReadableMode, ModuleListOptions, ModuleListRequest,
+        RemoteFallbackArgs, RemoteFallbackContext, StrongChecksumChoice, TransferTimeout,
+        parse_skip_compress_list, run_client_or_fallback,
         run_module_list_with_password_and_options, skip_compress_from_env,
     },
     fallback::{CLIENT_FALLBACK_ENV, FallbackOverride, fallback_override},
@@ -128,6 +128,10 @@ use rsync_meta::ChmodModifiers;
 use rsync_protocol::{ParseProtocolVersionErrorKind, ProtocolVersion};
 use time::{OffsetDateTime, format_description::FormatItem, macros::format_description};
 use users::{get_group_by_gid, get_group_by_name, get_user_by_name, get_user_by_uid, gid_t, uid_t};
+
+mod progress;
+
+pub(crate) use progress::*;
 
 /// Maximum exit code representable by a Unix process.
 const MAX_EXIT_CODE: i32 = u8::MAX as i32;
@@ -947,38 +951,6 @@ fn format_full_checksum(event: &ClientEvent) -> String {
         rendered.push_str(&format!("{byte:02x}"));
     }
     rendered
-}
-
-/// Parsed command produced by [`parse_args`].
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-enum ProgressMode {
-    PerFile,
-    Overall,
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-enum ProgressSetting {
-    Unspecified,
-    Disabled,
-    PerFile,
-    Overall,
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-enum NameOutputLevel {
-    Disabled,
-    UpdatedOnly,
-    UpdatedAndUnchanged,
-}
-
-impl ProgressSetting {
-    fn resolved(self) -> Option<ProgressMode> {
-        match self {
-            Self::PerFile => Some(ProgressMode::PerFile),
-            Self::Overall => Some(ProgressMode::Overall),
-            Self::Disabled | Self::Unspecified => None,
-        }
-    }
 }
 
 #[derive(Clone, Copy, Debug, Eq, PartialEq)]
@@ -4244,992 +4216,6 @@ where
             }
             error.exit_code()
         }
-    }
-}
-
-/// Emits verbose, statistics, and progress-oriented output derived from a [`ClientSummary`].
-struct LiveProgress<'a> {
-    writer: &'a mut dyn Write,
-    rendered: bool,
-    error: Option<io::Error>,
-    active_path: Option<PathBuf>,
-    line_active: bool,
-    mode: ProgressMode,
-    human_readable: HumanReadableMode,
-}
-
-impl<'a> LiveProgress<'a> {
-    fn new(
-        writer: &'a mut dyn Write,
-        mode: ProgressMode,
-        human_readable: HumanReadableMode,
-    ) -> Self {
-        Self {
-            writer,
-            rendered: false,
-            error: None,
-            active_path: None,
-            line_active: false,
-            mode,
-            human_readable,
-        }
-    }
-
-    fn rendered(&self) -> bool {
-        self.rendered
-    }
-
-    fn record_error(&mut self, error: io::Error) {
-        if self.error.is_none() {
-            self.error = Some(error);
-        }
-    }
-
-    fn finish(self) -> io::Result<()> {
-        if let Some(error) = self.error {
-            return Err(error);
-        }
-
-        if self.line_active {
-            writeln!(self.writer)?;
-        }
-
-        Ok(())
-    }
-}
-
-impl<'a> ClientProgressObserver for LiveProgress<'a> {
-    fn on_progress(&mut self, update: &ClientProgressUpdate) {
-        if self.error.is_some() {
-            return;
-        }
-
-        let total = update.total().max(update.index());
-        let remaining = total.saturating_sub(update.index());
-
-        let write_result = match self.mode {
-            ProgressMode::PerFile => (|| -> io::Result<()> {
-                let event = update.event();
-                let relative = event.relative_path();
-                let path_changed = self.active_path.as_deref() != Some(relative);
-
-                if path_changed {
-                    if self.line_active {
-                        writeln!(self.writer)?;
-                        self.line_active = false;
-                    }
-                    writeln!(self.writer, "{}", relative.display())?;
-                    self.active_path = Some(relative.to_path_buf());
-                }
-
-                let bytes = event.bytes_transferred();
-                let size_field =
-                    format!("{:>15}", format_progress_bytes(bytes, self.human_readable));
-                let percent = format_progress_percent(bytes, update.total_bytes());
-                let percent_field = format!("{:>4}", percent);
-                let rate_field = format!(
-                    "{:>12}",
-                    format_progress_rate(bytes, event.elapsed(), self.human_readable)
-                );
-                let elapsed_field = format!("{:>11}", format_progress_elapsed(event.elapsed()));
-                let xfr_index = update.index();
-
-                if self.line_active {
-                    write!(self.writer, "\r")?;
-                }
-
-                write!(
-                    self.writer,
-                    "{size_field} {percent_field} {rate_field} {elapsed_field} (xfr#{xfr_index}, to-chk={remaining}/{total})"
-                )?;
-
-                if update.is_final() {
-                    writeln!(self.writer)?;
-                    self.line_active = false;
-                    self.active_path = None;
-                } else {
-                    self.line_active = true;
-                }
-                Ok(())
-            })(),
-            ProgressMode::Overall => (|| -> io::Result<()> {
-                let bytes = update.overall_transferred();
-                let size_field =
-                    format!("{:>15}", format_progress_bytes(bytes, self.human_readable));
-                let percent_field = format!(
-                    "{:>4}",
-                    format_progress_percent(bytes, update.overall_total_bytes())
-                );
-                let rate_field = format!(
-                    "{:>12}",
-                    format_progress_rate(bytes, update.overall_elapsed(), self.human_readable)
-                );
-                let elapsed_field =
-                    format!("{:>11}", format_progress_elapsed(update.overall_elapsed()));
-                let xfr_index = update.index();
-
-                if self.line_active {
-                    write!(self.writer, "\r")?;
-                }
-
-                write!(
-                    self.writer,
-                    "{size_field} {percent_field} {rate_field} {elapsed_field} (xfr#{xfr_index}, to-chk={remaining}/{total})"
-                )?;
-
-                if update.remaining() == 0 && update.is_final() {
-                    writeln!(self.writer)?;
-                    self.line_active = false;
-                } else {
-                    self.line_active = true;
-                }
-                self.active_path = None;
-                Ok(())
-            })(),
-        };
-
-        match write_result {
-            Ok(()) => {
-                self.rendered = true;
-            }
-            Err(error) => self.record_error(error),
-        }
-    }
-}
-
-#[allow(clippy::too_many_arguments)]
-fn emit_transfer_summary(
-    summary: &ClientSummary,
-    verbosity: u8,
-    progress_mode: Option<ProgressMode>,
-    stats: bool,
-    progress_already_rendered: bool,
-    list_only: bool,
-    out_format: Option<&OutFormat>,
-    out_format_context: &OutFormatContext,
-    name_level: NameOutputLevel,
-    name_overridden: bool,
-    human_readable_mode: HumanReadableMode,
-    writer: &mut dyn Write,
-) -> io::Result<()> {
-    let events = summary.events();
-
-    if list_only {
-        let mut wrote_listing = false;
-        if !events.is_empty() {
-            emit_list_only(events, writer, human_readable_mode)?;
-            wrote_listing = true;
-        }
-
-        if stats {
-            if wrote_listing {
-                writeln!(writer)?;
-            }
-            emit_stats(summary, writer, human_readable_mode)?;
-        } else if verbosity > 0 {
-            if wrote_listing {
-                writeln!(writer)?;
-            }
-            emit_totals(summary, writer, human_readable_mode)?;
-        }
-
-        return Ok(());
-    }
-
-    let formatted_rendered = if let Some(format) = out_format {
-        if events.is_empty() {
-            false
-        } else {
-            emit_out_format(events, format, out_format_context, writer)?;
-            true
-        }
-    } else {
-        false
-    };
-
-    let progress_rendered = if progress_already_rendered {
-        true
-    } else if matches!(progress_mode, Some(ProgressMode::PerFile)) && !events.is_empty() {
-        emit_progress(events, writer, human_readable_mode)?
-    } else {
-        false
-    };
-
-    let name_enabled = !matches!(name_level, NameOutputLevel::Disabled);
-    let emit_verbose_listing = out_format.is_none()
-        && !events.is_empty()
-        && ((verbosity > 0
-            && (!name_overridden || name_enabled)
-            && (!progress_rendered || verbosity > 1))
-            || (verbosity == 0 && name_enabled));
-
-    if formatted_rendered && (emit_verbose_listing || stats || verbosity > 0) {
-        writeln!(writer)?;
-    }
-
-    if progress_rendered && (emit_verbose_listing || stats || verbosity > 0) {
-        writeln!(writer)?;
-    }
-
-    if emit_verbose_listing {
-        emit_verbose(
-            events,
-            verbosity,
-            name_level,
-            name_overridden,
-            human_readable_mode,
-            writer,
-        )?;
-        if stats {
-            writeln!(writer)?;
-        }
-    }
-
-    if stats {
-        emit_stats(summary, writer, human_readable_mode)?;
-    } else if verbosity > 0 || (verbosity == 0 && name_enabled) {
-        emit_totals(summary, writer, human_readable_mode)?;
-    }
-
-    Ok(())
-}
-
-fn emit_list_only<W: Write + ?Sized>(
-    events: &[ClientEvent],
-    stdout: &mut W,
-    human_readable: HumanReadableMode,
-) -> io::Result<()> {
-    for event in events {
-        if !list_only_event(event.kind()) {
-            continue;
-        }
-
-        if let Some(metadata) = event.metadata() {
-            let permissions = format_list_permissions(metadata);
-            let size = format_list_size(metadata.length(), human_readable);
-            let timestamp = format_list_timestamp(metadata.modified());
-            let mut rendered = event.relative_path().to_string_lossy().into_owned();
-            if metadata.kind().is_symlink() {
-                if let Some(target) = metadata.symlink_target() {
-                    rendered.push_str(" -> ");
-                    rendered.push_str(&target.to_string_lossy());
-                }
-            }
-
-            writeln!(stdout, "{permissions} {size} {timestamp} {rendered}")?;
-        } else {
-            let rendered = event.relative_path().to_string_lossy().into_owned();
-            writeln!(
-                stdout,
-                "?????????? {:>15} {} {rendered}",
-                "?",
-                format_list_timestamp(None),
-            )?;
-        }
-    }
-
-    Ok(())
-}
-
-fn list_only_event(kind: &ClientEventKind) -> bool {
-    matches!(
-        kind,
-        ClientEventKind::DataCopied
-            | ClientEventKind::MetadataReused
-            | ClientEventKind::HardLink
-            | ClientEventKind::SymlinkCopied
-            | ClientEventKind::FifoCopied
-            | ClientEventKind::DeviceCopied
-            | ClientEventKind::DirectoryCreated
-    )
-}
-
-fn format_list_permissions(metadata: &ClientEntryMetadata) -> String {
-    let type_char = match metadata.kind() {
-        ClientEntryKind::File => '-',
-        ClientEntryKind::Directory => 'd',
-        ClientEntryKind::Symlink => 'l',
-        ClientEntryKind::Fifo => 'p',
-        ClientEntryKind::CharDevice => 'c',
-        ClientEntryKind::BlockDevice => 'b',
-        ClientEntryKind::Socket => 's',
-        ClientEntryKind::Other => '?',
-    };
-
-    let mut symbols = ['-'; 10];
-    symbols[0] = type_char;
-
-    if let Some(mode) = metadata.mode() {
-        const PERMISSION_MASKS: [(usize, u32, char); 9] = [
-            (1, 0o400, 'r'),
-            (2, 0o200, 'w'),
-            (3, 0o100, 'x'),
-            (4, 0o040, 'r'),
-            (5, 0o020, 'w'),
-            (6, 0o010, 'x'),
-            (7, 0o004, 'r'),
-            (8, 0o002, 'w'),
-            (9, 0o001, 'x'),
-        ];
-
-        for &(index, mask, ch) in &PERMISSION_MASKS {
-            if mode & mask != 0 {
-                symbols[index] = ch;
-            }
-        }
-
-        if mode & 0o4000 != 0 {
-            symbols[3] = match symbols[3] {
-                'x' => 's',
-                '-' => 'S',
-                other => other,
-            };
-        }
-
-        if mode & 0o2000 != 0 {
-            symbols[6] = match symbols[6] {
-                'x' => 's',
-                '-' => 'S',
-                other => other,
-            };
-        }
-
-        if mode & 0o1000 != 0 {
-            symbols[9] = match symbols[9] {
-                'x' => 't',
-                '-' => 'T',
-                other => other,
-            };
-        }
-    }
-
-    symbols.iter().collect()
-}
-
-fn format_list_timestamp(modified: Option<SystemTime>) -> String {
-    if let Some(time) = modified {
-        if let Ok(datetime) = OffsetDateTime::from(time).format(LIST_TIMESTAMP_FORMAT) {
-            return datetime;
-        }
-    }
-    "1970/01/01 00:00:00".to_string()
-}
-
-fn format_list_size(size: u64, human_readable: HumanReadableMode) -> String {
-    let value = format_size(size, human_readable);
-    format!("{value:>15}")
-}
-
-/// Returns whether the provided event kind should be reflected in progress output.
-fn is_progress_event(kind: &ClientEventKind) -> bool {
-    kind.is_progress()
-}
-
-/// Formats a byte count using thousands separators when human-readable formatting is disabled. When
-/// enabled, the output uses decimal unit suffixes such as `K`, `M`, or `G` with two fractional
-/// digits. Combined mode includes the exact decimal value in parentheses when the two representations
-/// differ.
-fn format_progress_bytes(bytes: u64, human_readable: HumanReadableMode) -> String {
-    format_size(bytes, human_readable)
-}
-
-fn format_size(bytes: u64, human_readable: HumanReadableMode) -> String {
-    let decimal = format_decimal_bytes(bytes);
-    if !human_readable.is_enabled() {
-        return decimal;
-    }
-
-    let human = format_human_bytes(bytes);
-    if human_readable.includes_exact() && human != decimal {
-        format!("{human} ({decimal})")
-    } else {
-        human
-    }
-}
-
-fn format_decimal_bytes(bytes: u64) -> String {
-    let mut digits = bytes.to_string();
-    let mut groups = Vec::new();
-
-    while digits.len() > 3 {
-        let chunk = digits.split_off(digits.len() - 3);
-        groups.push(chunk);
-    }
-
-    groups.push(digits);
-    groups.reverse();
-    groups.join(",")
-}
-
-fn format_human_bytes(bytes: u64) -> String {
-    if bytes < 1_000 {
-        return bytes.to_string();
-    }
-
-    const UNITS: &[(&str, f64)] = &[
-        ("P", 1_000_000_000_000_000.0),
-        ("T", 1_000_000_000_000.0),
-        ("G", 1_000_000_000.0),
-        ("M", 1_000_000.0),
-        ("K", 1_000.0),
-    ];
-
-    let bytes_f64 = bytes as f64;
-    for (suffix, threshold) in UNITS {
-        if bytes_f64 >= *threshold {
-            let value = bytes_f64 / *threshold;
-            return format!("{value:.2}{suffix}");
-        }
-    }
-
-    bytes.to_string()
-}
-
-fn format_summary_rate(rate: f64, human_readable: HumanReadableMode) -> String {
-    let decimal = format!("{rate:.2}");
-    if !human_readable.is_enabled() {
-        return decimal;
-    }
-
-    let human = format_human_rate(rate);
-    if human_readable.includes_exact() && human != decimal {
-        format!("{human} ({decimal})")
-    } else {
-        human
-    }
-}
-
-fn format_human_rate(rate: f64) -> String {
-    if rate < 1_000.0 {
-        return format!("{rate:.2}");
-    }
-
-    const UNITS: &[(&str, f64)] = &[
-        ("P", 1_000_000_000_000_000.0),
-        ("T", 1_000_000_000_000.0),
-        ("G", 1_000_000_000.0),
-        ("M", 1_000_000.0),
-        ("K", 1_000.0),
-    ];
-
-    for (suffix, threshold) in UNITS {
-        if rate >= *threshold {
-            let value = rate / *threshold;
-            return format!("{value:.2}{suffix}");
-        }
-    }
-
-    format!("{rate:.2}")
-}
-
-struct VerboseRateDisplay {
-    primary: (String, &'static str),
-    secondary: Option<(String, &'static str)>,
-}
-
-fn format_verbose_rate(rate: f64, human_readable: HumanReadableMode) -> VerboseRateDisplay {
-    let decimal = format_verbose_rate_decimal(rate);
-    if !human_readable.is_enabled() {
-        return VerboseRateDisplay {
-            primary: decimal,
-            secondary: None,
-        };
-    }
-
-    let human = format_verbose_rate_human(rate);
-    if human_readable.includes_exact() && (human.0 != decimal.0 || human.1 != decimal.1) {
-        VerboseRateDisplay {
-            primary: human,
-            secondary: Some(decimal),
-        }
-    } else {
-        VerboseRateDisplay {
-            primary: human,
-            secondary: None,
-        }
-    }
-}
-
-fn format_verbose_rate_decimal(rate: f64) -> (String, &'static str) {
-    (format!("{rate:.1}"), "B/s")
-}
-
-fn format_verbose_rate_human(rate: f64) -> (String, &'static str) {
-    const UNITS: &[(&str, f64)] = &[
-        ("PB/s", 1_000_000_000_000_000.0),
-        ("TB/s", 1_000_000_000_000.0),
-        ("GB/s", 1_000_000_000.0),
-        ("MB/s", 1_000_000.0),
-        ("kB/s", 1_000.0),
-    ];
-
-    for (unit, threshold) in UNITS {
-        if rate >= *threshold {
-            let value = rate / *threshold;
-            return (format!("{value:.2}"), *unit);
-        }
-    }
-
-    (format!("{rate:.2}"), "B/s")
-}
-
-/// Formats a progress percentage, producing the upstream `??%` placeholder when totals are
-/// unavailable.
-fn format_progress_percent(bytes: u64, total: Option<u64>) -> String {
-    match total {
-        Some(total_bytes) if total_bytes > 0 => {
-            let capped = bytes.min(total_bytes);
-            let percent = (capped.saturating_mul(100)) / total_bytes;
-            format!("{percent}%")
-        }
-        Some(_) => "100%".to_string(),
-        None => "??%".to_string(),
-    }
-}
-
-/// Formats a transfer rate in the `kB/s`, `MB/s`, or `GB/s` ranges.
-fn format_progress_rate(
-    bytes: u64,
-    elapsed: Duration,
-    human_readable: HumanReadableMode,
-) -> String {
-    if bytes == 0 || elapsed.is_zero() {
-        return if human_readable.is_enabled() {
-            "0.00B/s".to_string()
-        } else {
-            "0.00kB/s".to_string()
-        };
-    }
-
-    let seconds = elapsed.as_secs_f64();
-    if seconds <= 0.0 {
-        return if human_readable.is_enabled() {
-            "0.00B/s".to_string()
-        } else {
-            "0.00kB/s".to_string()
-        };
-    }
-
-    let rate = bytes as f64 / seconds;
-    let decimal = format_progress_rate_decimal(rate);
-    if !human_readable.is_enabled() {
-        return decimal;
-    }
-
-    let human = format_progress_rate_human(rate);
-    if human_readable.includes_exact() && human != decimal {
-        format!("{human} ({decimal})")
-    } else {
-        human
-    }
-}
-
-fn format_progress_rate_decimal(rate: f64) -> String {
-    const KIB: f64 = 1024.0;
-    const MIB: f64 = KIB * 1024.0;
-    const GIB: f64 = MIB * 1024.0;
-
-    if rate >= GIB {
-        format!("{:.2}GB/s", rate / GIB)
-    } else if rate >= MIB {
-        format!("{:.2}MB/s", rate / MIB)
-    } else {
-        format!("{:.2}kB/s", rate / KIB)
-    }
-}
-
-fn format_progress_rate_human(rate: f64) -> String {
-    let display = format_verbose_rate_human(rate);
-    format!("{}{}", display.0, display.1)
-}
-
-/// Formats an elapsed duration as `H:MM:SS`, matching rsync's progress output.
-fn format_progress_elapsed(elapsed: Duration) -> String {
-    let total_seconds = elapsed.as_secs();
-    let hours = total_seconds / 3_600;
-    let minutes = (total_seconds % 3_600) / 60;
-    let seconds = total_seconds % 60;
-    format!("{hours}:{minutes:02}:{seconds:02}")
-}
-
-/// Renders progress lines for the provided transfer events.
-fn emit_progress<W: Write + ?Sized>(
-    events: &[ClientEvent],
-    stdout: &mut W,
-    human_readable: HumanReadableMode,
-) -> io::Result<bool> {
-    let progress_events: Vec<_> = events
-        .iter()
-        .filter(|event| is_progress_event(event.kind()))
-        .collect();
-
-    if progress_events.is_empty() {
-        return Ok(false);
-    }
-
-    let total = progress_events.len();
-
-    for (index, event) in progress_events.into_iter().enumerate() {
-        writeln!(stdout, "{}", event.relative_path().display())?;
-
-        let bytes = event.bytes_transferred();
-        let size_field = format!("{:>15}", format_progress_bytes(bytes, human_readable));
-        let percent_hint = match event.kind() {
-            ClientEventKind::DataCopied => event.total_bytes(),
-            _ => None,
-        };
-        let percent_field = format!("{:>4}", format_progress_percent(bytes, percent_hint));
-        let rate_field = format!(
-            "{:>12}",
-            format_progress_rate(bytes, event.elapsed(), human_readable)
-        );
-        let elapsed_field = format!("{:>11}", format_progress_elapsed(event.elapsed()));
-        let remaining = total - index - 1;
-        let xfr_index = index + 1;
-
-        writeln!(
-            stdout,
-            "{size_field} {percent_field} {rate_field} {elapsed_field} (xfr#{xfr_index}, to-chk={remaining}/{total})"
-        )?;
-    }
-
-    Ok(true)
-}
-
-/// Emits a statistics summary mirroring the subset of counters supported by the local engine.
-fn emit_stats<W: Write + ?Sized>(
-    summary: &ClientSummary,
-    stdout: &mut W,
-    human_readable: HumanReadableMode,
-) -> io::Result<()> {
-    let files = summary.files_copied();
-    let files_total = summary.regular_files_total();
-    let directories = summary.directories_created();
-    let directories_total = summary.directories_total();
-    let symlinks = summary.symlinks_copied();
-    let symlinks_total = summary.symlinks_total();
-    let devices = summary.devices_created();
-    let devices_total = summary.devices_total();
-    let fifos = summary.fifos_created();
-    let fifos_total = summary.fifos_total();
-    let deleted = summary.items_deleted();
-    let literal_bytes = summary.bytes_copied();
-    let transferred_size = summary.transferred_file_size();
-    let bytes_sent = summary.bytes_sent();
-    let bytes_received = summary.bytes_received();
-    let total_size = summary.total_source_bytes();
-    let matched_bytes = summary.matched_bytes();
-    let file_list_size = summary.file_list_size();
-    let file_list_generation = summary.file_list_generation_time().as_secs_f64();
-    let file_list_transfer = summary.file_list_transfer_time().as_secs_f64();
-
-    let special_total = devices_total.saturating_add(fifos_total);
-    let special_created = devices.saturating_add(fifos);
-    let total_entries = files_total
-        .saturating_add(directories_total)
-        .saturating_add(symlinks_total)
-        .saturating_add(special_total);
-    let created_total = files
-        .saturating_add(directories)
-        .saturating_add(symlinks)
-        .saturating_add(special_created);
-
-    let files_breakdown = format_stat_categories(&[
-        ("reg", files_total),
-        ("dir", directories_total),
-        ("link", symlinks_total),
-        ("special", special_total),
-    ]);
-    let created_breakdown = format_stat_categories(&[
-        ("reg", files),
-        ("dir", directories),
-        ("link", symlinks),
-        ("special", special_created),
-    ]);
-
-    let total_size_display = format_size(total_size, human_readable);
-    let transferred_size_display = format_size(transferred_size, human_readable);
-    let literal_bytes_display = format_size(literal_bytes, human_readable);
-    let matched_bytes_display = format_size(matched_bytes, human_readable);
-    let file_list_size_display = format_size(file_list_size, human_readable);
-    let bytes_sent_display = format_size(bytes_sent, human_readable);
-    let bytes_received_display = format_size(bytes_received, human_readable);
-
-    writeln!(stdout, "Number of files: {total_entries}{files_breakdown}")?;
-    writeln!(
-        stdout,
-        "Number of created files: {created_total}{created_breakdown}"
-    )?;
-    writeln!(stdout, "Number of deleted files: {deleted}")?;
-    writeln!(stdout, "Number of regular files transferred: {files}")?;
-    writeln!(stdout, "Total file size: {total_size_display} bytes")?;
-    writeln!(
-        stdout,
-        "Total transferred file size: {transferred_size_display} bytes"
-    )?;
-    writeln!(stdout, "Literal data: {literal_bytes_display} bytes")?;
-    writeln!(stdout, "Matched data: {matched_bytes_display} bytes")?;
-    writeln!(stdout, "File list size: {file_list_size_display}")?;
-    writeln!(
-        stdout,
-        "File list generation time: {file_list_generation:.3} seconds"
-    )?;
-    writeln!(
-        stdout,
-        "File list transfer time: {file_list_transfer:.3} seconds"
-    )?;
-    writeln!(stdout, "Total bytes sent: {bytes_sent_display}")?;
-    writeln!(stdout, "Total bytes received: {bytes_received_display}")?;
-    writeln!(stdout)?;
-
-    emit_totals(summary, stdout, human_readable)
-}
-
-fn format_stat_categories(categories: &[(&str, u64)]) -> String {
-    let parts: Vec<String> = categories
-        .iter()
-        .filter(|(_, count)| *count > 0)
-        .map(|(label, count)| format!("{label}: {count}"))
-        .collect();
-    if parts.is_empty() {
-        String::new()
-    } else {
-        format!(" ({})", parts.join(", "))
-    }
-}
-
-/// Emits the summary lines reported by verbose transfers.
-fn emit_totals<W: Write + ?Sized>(
-    summary: &ClientSummary,
-    stdout: &mut W,
-    human_readable: HumanReadableMode,
-) -> io::Result<()> {
-    let sent = summary.bytes_sent();
-    let received = summary.bytes_received();
-    let total_size = summary.total_source_bytes();
-    let elapsed = summary.total_elapsed();
-    let seconds = elapsed.as_secs_f64();
-    let rate = if seconds > 0.0 {
-        (sent + received) as f64 / seconds
-    } else {
-        0.0
-    };
-    let transmitted = sent.saturating_add(received);
-    let speedup = if transmitted > 0 {
-        total_size as f64 / transmitted as f64
-    } else {
-        0.0
-    };
-
-    let sent_display = format_size(sent, human_readable);
-    let received_display = format_size(received, human_readable);
-    let rate_display = format_summary_rate(rate, human_readable);
-    let total_size_display = format_size(total_size, human_readable);
-
-    writeln!(
-        stdout,
-        "sent {sent_display} bytes  received {received_display} bytes  {rate_display} bytes/sec"
-    )?;
-    writeln!(
-        stdout,
-        "total size is {total_size_display}  speedup is {speedup:.2}"
-    )
-}
-
-/// Renders verbose listings for the provided transfer events.
-fn emit_verbose<W: Write + ?Sized>(
-    events: &[ClientEvent],
-    verbosity: u8,
-    name_level: NameOutputLevel,
-    name_overridden: bool,
-    human_readable: HumanReadableMode,
-    stdout: &mut W,
-) -> io::Result<()> {
-    if matches!(name_level, NameOutputLevel::Disabled) && (verbosity == 0 || name_overridden) {
-        return Ok(());
-    }
-
-    for event in events {
-        let kind = event.kind();
-        let include_for_name = event_matches_name_level(event, name_level);
-
-        if verbosity == 0 {
-            if !include_for_name {
-                continue;
-            }
-
-            let mut rendered = event.relative_path().to_string_lossy().into_owned();
-            if matches!(kind, ClientEventKind::SymlinkCopied) {
-                if let Some(metadata) = event.metadata() {
-                    if let Some(target) = metadata.symlink_target() {
-                        rendered.push_str(" -> ");
-                        rendered.push_str(&target.to_string_lossy());
-                    }
-                }
-            }
-            writeln!(stdout, "{rendered}")?;
-            continue;
-        }
-
-        if name_overridden && !include_for_name {
-            continue;
-        }
-
-        match kind {
-            ClientEventKind::SkippedExisting => {
-                writeln!(
-                    stdout,
-                    "skipping existing file \"{}\"",
-                    event.relative_path().display()
-                )?;
-                continue;
-            }
-            ClientEventKind::SkippedNewerDestination => {
-                writeln!(
-                    stdout,
-                    "skipping newer destination file \"{}\"",
-                    event.relative_path().display()
-                )?;
-                continue;
-            }
-            ClientEventKind::SkippedNonRegular => {
-                writeln!(
-                    stdout,
-                    "skipping non-regular file \"{}\"",
-                    event.relative_path().display()
-                )?;
-                continue;
-            }
-            ClientEventKind::SkippedUnsafeSymlink => {
-                let mut rendered = format!(
-                    "ignoring unsafe symlink \"{}\"",
-                    event.relative_path().display()
-                );
-                if let Some(metadata) = event.metadata() {
-                    if let Some(target) = metadata.symlink_target() {
-                        rendered.push_str(" -> ");
-                        rendered.push_str(&target.to_string_lossy());
-                    }
-                }
-                writeln!(stdout, "{rendered}")?;
-                continue;
-            }
-            ClientEventKind::SkippedMountPoint => {
-                writeln!(
-                    stdout,
-                    "skipping mount point \"{}\"",
-                    event.relative_path().display()
-                )?;
-                continue;
-            }
-            _ => {}
-        }
-
-        let mut rendered = event.relative_path().to_string_lossy().into_owned();
-        if matches!(kind, ClientEventKind::SymlinkCopied) {
-            if let Some(metadata) = event.metadata() {
-                if let Some(target) = metadata.symlink_target() {
-                    rendered.push_str(" -> ");
-                    rendered.push_str(&target.to_string_lossy());
-                }
-            }
-        }
-
-        if verbosity == 1 {
-            writeln!(stdout, "{rendered}")?;
-            continue;
-        }
-
-        let descriptor = describe_event_kind(kind);
-        let bytes = event.bytes_transferred();
-        if bytes > 0 {
-            let size_text = format_size(bytes, human_readable);
-            if let Some(rate) = compute_rate(bytes, event.elapsed()) {
-                let rate_display = format_verbose_rate(rate, human_readable);
-                if let Some((secondary_value, secondary_unit)) = rate_display.secondary {
-                    writeln!(
-                        stdout,
-                        "{descriptor}: {rendered} ({size_text} bytes, {} {} ({} {}))",
-                        rate_display.primary.0,
-                        rate_display.primary.1,
-                        secondary_value,
-                        secondary_unit
-                    )?;
-                } else {
-                    writeln!(
-                        stdout,
-                        "{descriptor}: {rendered} ({size_text} bytes, {} {})",
-                        rate_display.primary.0, rate_display.primary.1
-                    )?;
-                }
-            } else {
-                writeln!(stdout, "{descriptor}: {rendered} ({size_text} bytes)")?;
-            }
-        } else {
-            writeln!(stdout, "{descriptor}: {rendered}")?;
-        }
-    }
-    Ok(())
-}
-
-fn event_matches_name_level(event: &ClientEvent, level: NameOutputLevel) -> bool {
-    match level {
-        NameOutputLevel::Disabled => false,
-        NameOutputLevel::UpdatedOnly => matches!(
-            event.kind(),
-            ClientEventKind::DataCopied
-                | ClientEventKind::HardLink
-                | ClientEventKind::SymlinkCopied
-                | ClientEventKind::FifoCopied
-                | ClientEventKind::DeviceCopied
-                | ClientEventKind::DirectoryCreated
-                | ClientEventKind::SourceRemoved
-        ),
-        NameOutputLevel::UpdatedAndUnchanged => matches!(
-            event.kind(),
-            ClientEventKind::DataCopied
-                | ClientEventKind::MetadataReused
-                | ClientEventKind::HardLink
-                | ClientEventKind::SymlinkCopied
-                | ClientEventKind::FifoCopied
-                | ClientEventKind::DeviceCopied
-                | ClientEventKind::DirectoryCreated
-                | ClientEventKind::SourceRemoved
-        ),
-    }
-}
-
-/// Maps an event kind to a human-readable description.
-fn describe_event_kind(kind: &ClientEventKind) -> &'static str {
-    match kind {
-        ClientEventKind::DataCopied => "copied",
-        ClientEventKind::MetadataReused => "metadata reused",
-        ClientEventKind::HardLink => "hard link",
-        ClientEventKind::SymlinkCopied => "symlink",
-        ClientEventKind::FifoCopied => "fifo",
-        ClientEventKind::DeviceCopied => "device",
-        ClientEventKind::DirectoryCreated => "directory",
-        ClientEventKind::SkippedExisting => "skipped existing file",
-        ClientEventKind::SkippedNonRegular => "skipped non-regular file",
-        ClientEventKind::SkippedUnsafeSymlink => "skipped unsafe symlink",
-        ClientEventKind::SkippedMountPoint => "skipped mount point",
-        ClientEventKind::SkippedNewerDestination => "skipped newer destination file",
-        ClientEventKind::EntryDeleted => "deleted",
-        ClientEventKind::SourceRemoved => "source removed",
-    }
-}
-
-/// Computes the throughput in bytes per second for the provided measurements.
-fn compute_rate(bytes: u64, elapsed: Duration) -> Option<f64> {
-    if elapsed.is_zero() {
-        return None;
-    }
-
-    let seconds = elapsed.as_secs_f64();
-    if seconds <= 0.0 {
-        None
-    } else {
-        Some(bytes as f64 / seconds)
     }
 }
